C: various fixes (including 32-bit)

+++ lib/hash-utility.c
@@ -202,7 +202,7 @@ int rm_hasher_main(int argc, const char **argv) {
     rm_digest_enable_sse(TRUE);
 #endif
 
-    int buf_size = (g_strv_length(tag.paths) + 1) * sizeof(RmDigest *);
+    gsize buf_size = (g_strv_length(tag.paths) + 1) * sizeof(RmDigest *);
     tag.read_succesful = g_slice_alloc0(buf_size);
 
     if(tag.print_in_order) {
